api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
scikit-learn
xgboost
joblib
boto3
prometheus-fastapi-instrumentator
python-dotenv
